refactor: global.configuration -> configuration
