+++ src/main/java/com/avalon/avalonchat/configuration/KeyValueStoreConfiguration.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.global.configuration;
+package com.avalon.avalonchat.configuration;
 
 import static java.nio.charset.StandardCharsets.*;
