+++ src/main/java/com/avalon/avalonchat/configuration/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.global.configuration.jwt;
+package com.avalon.avalonchat.configuration.jwt;
 
 import java.io.IOException;
 import java.util.Map;
