+++ src/main/java/com/avalon/avalonchat/configuration/jwt/JwtConfigProperties.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.global.configuration.jwt;
+package com.avalon.avalonchat.configuration.jwt;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.ConstructorBinding;
