+++ src/test/java/com/avalon/avalonchat/global/configuration/jwt/JwtConfigPropertiesTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import com.avalon.avalonchat.configuration.jwt.JwtConfigProperties;
+
 @SpringBootTest
 class JwtConfigPropertiesTest {
