+++ src/test/java/com/avalon/avalonchat/testsupport/base/BaseControllerTest.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.avalon.avalonchat.global.configuration.JacksonConfiguration;
+import com.avalon.avalonchat.configuration.JacksonConfiguration;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
